Introduced indeterminately failing BreakyBotSpec.
The possible outcomes of the final two BreakyBotSpec tests are the permutations of the set {fail,success}. This is either a result of a race condition in Akka, or my own silly mistake.

Also, running `test` repeatedly in the sbt console causes a PermGen exception, or a OOM: fun stuff.